fix: create routers folder because more code needed for models apis now (very confusing) + download persistency for the frontend
