+++ data/cosmic/statistic/03-2026.csv
@@ -1,2 +1,2 @@
 user_id,email,start_date,last_date,average_token_length,query_count
-1,smanileee@gmail.com,"10-03-2026,00:14:53","10-03-2026,21:19:52",11.96774193548387,31
+1,smanileee@gmail.com,"10-03-2026,00:14:53","22-03-2026,11:07:22",11.65625,32
