--- scripts/configs/config_updated.yaml
@@ -1,15 +1,0 @@
-chess:
-  stockfish_path: ''
-doc_directory: ''
-document_path: []
-is_quantized: false
-llm_name: ollama:qwen2.5:7b
-query_analyser:
-  is_quantized: false
-  llm_name: ollama:qwen2.5:7b
-rag:
-  retrieve_score_threshold: 0.7
-  topk: 10
-  vector_db_path: backend/data/vector_db_cosmic
-seed: 0
-service: -1
